Fix a bug in testing.empty

+++ cupy/testing/helper.py
@@ -1190,7 +1190,7 @@ def shaped_random(shape, xp=cupy, dtype=numpy.float32, scale=10, seed=0):
 
 
 def empty(xp=cupy, dtype=numpy.float32):
-    return xp.zeros((0,))
+    return xp.zeros((0,), dtype=dtype)
 
 
 class NumpyError(object):
